Move Thenable class to it's own file (no changes)
